+++ tests/number/isEven.js
@@ -1,0 +1,12 @@
+import isEven from '../../src/number/isEven'
+import test from 'tape'
+
+test('Handles checking for even numbers', t => {
+  t.true(isEven(2))
+  t.true(isEven(12))
+  t.true(isEven(4))
+  t.false(isEven(1))
+  t.false(isEven(-1))
+  t.false(isEven(NaN))
+  t.end()
+})