+++ tests/number/isOdd.js
@@ -1,0 +1,12 @@
+import isOdd from '../../src/number/isOdd'
+import test from 'tape'
+
+test('Checks numbers for odds', t => {
+  t.true(isOdd(1))
+  t.true(isOdd(3))
+  t.true(isOdd(15))
+  t.false(isOdd(2))
+  t.false(isOdd(4))
+  t.false(isOdd(NaN))
+  t.end()
+})